% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspecDB-package.R
\docType{data}
\name{C_cal}
\alias{C_cal}
\title{C_cal: NIR calibration for Carbon}
\format{An object of class \code{PLScalibration}.}
\description{
NIR calibration for Carbon using version 1 of plantspecDB, excluding
litter and bryophytes.
}
\examples{

}
\references{
\url{http://www.nutnet.umn.edu/}
}
\keyword{calibrations}
