% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspecDB-package.R
\docType{data}
\name{K_cal}
\alias{K_cal}
\title{K_cal: NIR calibration for Potassium}
\format{An object of class \code{PLScalibration}.}
\description{
NIR calibration for Potassium using version 1 of plantspecDB, excluding
litter and bryophytes.
}
\examples{

}
\references{
\url{http://www.nutnet.umn.edu/}
}
\keyword{calibrations}
