\name{leaf.spec.data}
\alias{leaf.spec.data}
\docType{data}
\title{
Wet lab chemistry for global calibration dataset.
}
\description{
Wet lab chemistry (a \code{data.frame}) from samples ground on an UDY belt drive mill.
}
\usage{data(leaf.spec.data)}
\format{
  A \code{data.frame} containing wet lab chemistry data and relevant metadata for the scans.

}

\references{
\url{http://www.nutnet.umn.edu/}

}
\examples{
data(leaf.spec.data)
}
\keyword{datasets}
