\name{leaf.spec.DB-package}
\alias{leaf.spec.DB-package}
\alias{leaf.spec.DB}
\docType{package}
\title{
NIR Calibration data for the 'leaf.spec' package
}
\description{
This R package is the companion package to \code{leaf.spec} and provides access to a global calibration dataset for leaf chemistry in herbaceous plants developed from the nutrient network experiment.
}

\author{
Daniel M. Griffith <griffith.dan@gmail.com>;
T. Michael Anderson <anderstm@wfu.edu>
}
\references{
\url{http://www.nutnet.umn.edu/}
}
\keyword{package}
\seealso{
}
\examples{
data(leaf.spec.spectra)
data(leaf.spec.data)
data(C_cal_23SEP2015)
data(N_cal_23SEP2015)
data(P_cal_23SEP2015)
data(K_cal_23SEP2015)
}
