\name{leaf.spec.spectra}
\alias{leaf.spec.spectra}
\docType{data}
\title{
NIR scans for global calibration dataset.
}
\description{
NIR scans (a \code{spectra.matrix}) from samples ground on an UDY belt drive mill and scanned on a Bruker MPA.
}
\usage{data(leaf.spec.spectra)}
\format{
  A large \code{spectra.matrix}.

}

\references{
\url{http://www.nutnet.umn.edu/}

}
\examples{
data(leaf.spec.spectra)
}
\keyword{datasets}
