% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspecDB-package.R
\docType{data}
\name{N_cal}
\alias{N_cal}
\title{N_cal: NIR calibration for Nitrogen}
\format{An object of class \code{PLScalibration}.}
\description{
NIR calibration for Nitrogen using version 1 of plantspecDB, excluding
litter and bryophytes.
}
\examples{

}
\references{
\url{http://www.nutnet.umn.edu/}
}
\keyword{calibrations}
