% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspecDB-package.R
\docType{data}
\name{P_cal}
\alias{P_cal}
\title{P_cal: NIR calibration for Phosphorus}
\format{An object of class \code{PLScalibration}.}
\description{
NIR calibration for Phosphorus using version 1 of plantspecDB, excluding
litter and bryophytes.
}
\examples{

}
\references{
\url{http://www.nutnet.umn.edu/}
}
\keyword{calibrations}
