% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspecDB-package.R
\docType{data}
\name{plantspec.data}
\alias{plantspec.data}
\title{Wet lab chemistry for global calibration dataset.}
\format{A \code{data.frame} containing wet lab chemistry data and relevant
metadata for the scans.}
\description{
Wet lab chemistry (a \code{data.frame}) from samples ground on an UDY belt
drive mill.
}
\examples{

data(plantspec.data)

}
\references{
\url{http://www.nutnet.umn.edu/}
}
\keyword{datasets}
