% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspecDB-package.R
\docType{data}
\name{plantspec.spectra}
\alias{plantspec.spectra}
\title{NIR scans for global calibration dataset.}
\format{A large \code{spectra.matrix}.}
\description{
NIR scans (a \code{spectra.matrix}) from samples ground on an UDY belt drive
mill and scanned on a Bruker MPA.
}
\examples{

data(plantspec.spectra)

}
\references{
\url{http://www.nutnet.umn.edu/}
}
\keyword{datasets}
