% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspecDB-package.R
\docType{package}
\name{plantspecDB-package}
\alias{plantspecDB-package}
\alias{plantspecDB}
\title{NIR Calibration data for the 'plantspec' package}
\description{
This R package is the companion package to \code{plantspec} and provides
access to a global calibration dataset for leaf chemistry in herbaceous
plants developed from the nutrient network experiment.
}
\examples{

data(plantspec.spectra)
data(plantspec.data)
data(C_cal)
data(N_cal)
data(P_cal)
data(K_cal)

}
\references{
\url{http://www.nutnet.umn.edu/}
}
\author{
Daniel M. Griffith <griffith.dan@gmail.com>; T. Michael Anderson
<anderstm@wfu.edu>
}
\keyword{package}
